requirements.txt
fastapi
uvicorn
pydantic
pydantic-ai
openai
httpx
supabase
python-dotenv
setuptools>=78.1.1
pytest
pytest-asyncio
stripe
python-jose[cryptography]
passlib[bcrypt]

